Sample data added and GeoSpatial Page improved
